vendor/cmd/go/internal/vgo: refer to GOPATH when no module is found
When FindModulePath gives up its search for a module in the current
directory, it complains with the message

"vgo: cannot determine module path for legacy source directory <dir>
(outside GOROOT, no import comments)"

Since FindModulePath is really searching through GOPATH, change the
reference in the error message to match.

While here, delete the word "legacy". It is confusing in the common
case of simply running vgo in the wrong directory.

Change-Id: Iba2b3e15233c373e978bcff43c98530fc74b9a50
Reviewed-on: https://go-review.googlesource.com/97195
Reviewed-by: Russ Cox <rsc@golang.org>