cmd/go/internal/modfetch: allow branch name as rev identifier
If someone runs "vgo get x@rev" or adds "require x rev" to go.mod
and rev is a commit hash or git tag, vgo automatically replaces it
with the pseudo-version for the underlying commit.
This CL makes vgo do the same for branch names:
the head commit of the branch at that moment is recorded.
This allows running

	vgo get github.com/gobuffalo/buffalo@development

to sync with the current buffalo development branch.

Fixes golang/go#24045.

Change-Id: I8a58ed3a574e5ee839d37ad452436f80ac8081dc
Reviewed-on: https://go-review.googlesource.com/105216
Reviewed-by: Bryan C. Mills <bcmills@google.com>