cmd/go/internal/vgo: speed up go.mod parsing
CL 106799 changed the mod file parser to be call
the fixer function for every module, version pair,
not just the ones with non-canonical semantic versions.
Of course, we don't want to hit the network for every line
of every go.mod, when most of them are fine.
Skip the network for the ones that are syntactically OK,
including proper matching between the module path and
major version.

For golang/go#24369.

Change-Id: Id101d8f3c10bccde8a755ae734dbacf4d0a36f8d
Reviewed-on: https://go-review.googlesource.com/107655
Run-TryBot: Russ Cox <rsc@golang.org>
Reviewed-by: Bryan C. Mills <bcmills@google.com>