cmd/go/internal/vgo: allow dotless module paths if replaced
In general a module path's first element should be a domain name,
or else we can't resolve it to find out how to download it.
But if a module is named and has a replace directive giving
a local directory where it can be found, we need not enforce
the leading-dot rule.

This test also serves to test that the go.mod version fixer does
not try to resolve the module path when it's well formed.

Fixes golang/go#24100.

Change-Id: I5cf3b82289db1f29b583f41fcf9f2509c7fc442a
Reviewed-on: https://go-review.googlesource.com/107661
Run-TryBot: Russ Cox <rsc@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Bryan C. Mills <bcmills@google.com>