x/vgo: fix parsing of go.mod with packages from gopkg.in
This addresses an issue where Parsing of `go.mod` with modules from
`gopkg.in` fails. The `go.mod` file is produced automatically, so the
first time something like `vgo build` is run (before `go.mod` is
created), it succeeds, but the second time it is run, it fails.

This fixes the latest issues described in
https://github.com/golang/go/issues/24099

The funcs `parseGopkgIn` and `dotV` were pulled in (and slightly
modified/simplified) from `vendor/cmd/go/internal/modfetch/gopkgin.go`.

I'm sure that this is not the best way to handle the special case of
`gopkg.in` (especially compared to the effort put into creating the
`VersionFixer`), but it's a starting point that has test cases that
prove the autogenerated `go.mod` file can't be parsed.

Fixes #24099

Signed-off-by: Joshua Rubin <joshua@rubixconsulting.com>