cmd/go/internal/modfetch/gitrepo: Support "dumb http" transport
The "dumb http" git transport is useful to host git repositories on a
static HTTP(s) server. Unfortunately, using git fetch --depth against
them is not supported, and fails with the error:

  fatal: dumb http transport does not support shallow capabilities

vgo currently fails on such repositories because of this.

This patch makes vgo identify this error, and retry the fetch without
--depth.

Change-Id: Id19396288f785f1c1f588e4fc1ff99eebc58bad9
Reviewed-on: https://go-review.googlesource.com/114820
Reviewed-by: Russ Cox <rsc@golang.org>