cmd/go/internal/modfetch/codehost: raise maximum repo size limit
This limit applies not just to the compressed zip file but also
to the total of the uncompressed contents.

github.com/Azure/azure-sdk-for-go has 116,335,208 bytes
of Go source files today. That exceeds the previous limit of
100 MB (104,857,600 bytes). Raise the limit.

For Azure/azure-sdk-for-go#1969.

Change-Id: Ibcc3b7b3fe375e2e17cc16210542b1d06293ba08
Reviewed-on: https://go-review.googlesource.com/116355
Run-TryBot: Russ Cox <rsc@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>
Reviewed-by: Bryan C. Mills <bcmills@google.com>