vendor/cmd/go/internal/modfetch: remove webtest usage in modfetch_test
The webtest hook no longer has any effect on these tests as of CL 107657.
That means the modfetch test is no longer hermetic (it actually talks
to GitHub), but a non-hermetic test is better than no test.

Keep webtest for future use: modfetch still uses web2 in a couple of places.

There are still a few failing tests in modfetch_test.
Those will be fixed in a later change.

Updates golang/go#25742.

Change-Id: Ib50838885fe9dc462ee9b03fcd9e29c1782e8dd6
Reviewed-on: https://go-review.googlesource.com/116835
Reviewed-by: Russ Cox <rsc@golang.org>