cmd/go/internal/vgo: skip not-yet-downloaded module in verify
Verify is meant to check what's on the local system.
If we haven't downloaded a module yet, don't download it just to check it.
If we do need to download it later, it will be checked then.

Fixes golang/go#24067.

Change-Id: Ic8dd83fac758404ff6d914447e077b2c54284136
Reviewed-on: https://go-review.googlesource.com/117935
Run-TryBot: Russ Cox <rsc@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Bryan C. Mills <bcmills@google.com>