cmd/go/internal/vgo: add -getmode=local, -getmode=vendor build flag
The new build flag -getmode=mode changes the way that vgo
resolves references to modules outside the target module.

The default (no -getmode flag) is to download and cache as usual.

With -getmode=local, the local cache is still consulted for modules,
but no network lookups are permitted.

With -getmode==vendor, only the vendor directory is used,
bypassing both the network and the local cache.
In this mode most of the interesting vgo behavior is inaccessible:
there is no world beyond the current module and its vendor directory.

I'm not thrilled with the name -getmode but don't have a better
name yet. It will probably change before the release - this CL is
about implementing the behavior.

Fixes golang/go#25073.

Change-Id: Ic5273f9f2d0263b49540401c4f554190702df099
Reviewed-on: https://go-review.googlesource.com/118316
Reviewed-by: Bryan C. Mills <bcmills@google.com>