cmd/go/internal/list: add module support
This CL changes the clumsy hard-coded "list -m" and "list -m -u"
with a real template-based implementation analogous to what we
do for listing packages. (In fact it reuses most of the package
implementation.) See the documentation for more details.

The main visible changes compared to the original vgo are:

- "vgo list -m" is now "vgo list -m all".
- "vgo list -m -u" is now "vgo list -m -u all".
- The default output is no longer a fixed-width-aligned table.
- The output no longer has COLUMN HEADERS.
- "vgo list -m -json [all]" now works.

Fixes golang/go#24347.
Fixes golang/go#25360.

Change-Id: Ie6a677875abca876ae334c8554298d3c56076b94
Reviewed-on: https://go-review.googlesource.com/120198
Reviewed-by: Bryan C. Mills <bcmills@google.com>