cmd/go/internal/modfetch: push codehost.Repo locking down one level
The modfetch.Repo needs to be safe for use by multiple goroutines,
but the codehost.Repo (a source code repository) was not,
so we added locking around the codehost.Repo methods.
But we might reasonably have multiple modfetch.Repos (modules)
drawing from a single codehost.Repo, especially if we cache
codehost.Repo construction. So push the locking further down.

After this CL, it would still be nice to add locking in package
codehost on a per-directory basis, so that codehost.Run calls
on a given directory are, or at least can be, serialized,
even when they are being invoked from different processes.
To do that we need to work out the right file locking on each
system, which we'll put off for now. (TODO added.)

(Git is already doing file locking to make access from multiple
processes safe at the level of individual commands.)

Change-Id: I9ce4ecf8f8b49cabea50b8663e3cee60ff95e7c9
Reviewed-on: https://go-review.googlesource.com/119775
Reviewed-by: Bryan C. Mills <bcmills@google.com>