cmd/go/internal/modfetch: always create and update go.sum
In the original vgo prototype, the go.modverify (now go.sum)
feature was opt-in: if you created the file, it was updated and
checked, but if the file didn't exist, vgo didn't create it.

After this CL, vgo will create the go.sum file automatically
(once there is a dependency to record). There is no way to
opt out.

If this turns out to be a problem for users for some reason,
we will change it back to being opt-in. But now is the time to
experiment.

Fixes golang/go#25525.

Change-Id: I1d1ee68da329d60cb8e2c8d76de96171d69f8e33
Reviewed-on: https://go-review.googlesource.com/121302
Reviewed-by: Bryan C. Mills <bcmills@google.com>