cmd/go/internal/vgo: prefer module path from Godeps/Godeps.json over GOPATH
The entry in Godeps/Godeps.json is explicitly what we're looking for:
the import path of the root directory.

The logic about deriving a module path from the current directory's
location in GOPATH should usually match, but it's less a direct statement:
for example, maybe a fork has been checked out.

If the two disagree, prefer Godeps/Godeps.json's explict statement.

Change-Id: I439684f5d34eeaf41419bbdae82139ba79498f6c
Reviewed-on: https://go-review.googlesource.com/121858
Reviewed-by: Bryan C. Mills <bcmills@google.com>