cmd/go/internal/modfile: add new requires to end of file
The old behavior was to add new requires to the first require statement in the file.

If there's more than one statement, though, it's more convenient for the first N-1
to be the manually maintained ones and then have the (possibly large) auto-maintained
block at the end of the file, instead of putting the manually maintained statements
after a giant auto-maintained one.

Suggested by Bryan Mills in an earlier CL, and works better for some real use cases.

Change-Id: I17a1704ca459c8cb5536dc23997bbacc3cd997eb
Reviewed-on: https://go-review.googlesource.com/121859
Reviewed-by: Bryan C. Mills <bcmills@google.com>