cmd/go: eliminate mentions of "vgo"
"vgo" is the name for the prototype/demo of Go module support.
What's merged into the Go toolchain needs to be "module support",
not a new tool ("vgo").

In preparation for that merge, remove uses of the name "vgo".
The tool will be just "go" again soon.

A few mentions of vgo remain, most notably in test repository names.
Those are OK to keep.

Change-Id: I7529860b430e9ce4b3e17f1bbb002df2f3ca76d6
Reviewed-on: https://go-review.googlesource.com/122264
Reviewed-by: Bryan C. Mills <bcmills@google.com>