cmd/go/internal/modfile: allow replace without old version
Replace directives to date have looked like

	replace old v1.2.3 => new v1.4.5

As pointed out in golang/go#24825, this can result in fragile go.mod
where the replacement stops kicking in after an upgrade that
affects the selected old version.

The fact that the replacement specifies exactly which version
it is capable of replacing might be a feature, but in other contexts
it's just a hassle. Allow not specifying the version, in which case
the replacement applies to all versions of that module.

Fixes golang/go#24825.

Change-Id: Iad97dde25c06fcba3c1181d650f9fa5e95f83157
Reviewed-on: https://go-review.googlesource.com/122400
Run-TryBot: Russ Cox <rsc@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Bryan C. Mills <bcmills@google.com>