cmd/go/internal/modfetch: make source cache directories unwritable
When we run tests of dependency packages,
we run them in their source directories
(as we do for all tests).

Discourage modification of the source directories
by marking them unwritable. This will undoubtedly
break some tests that write temporary files to the
current directory and then clean up after themselves.
That's OK: those tests should be fixed.
(We have a bunch of those ourselves, to be sure.)

Change-Id: Ie0e6713c0121a359b2f7153fe6cf27eac381b931
Reviewed-on: https://go-review.googlesource.com/122401
Reviewed-by: Bryan C. Mills <bcmills@google.com>