cmd/go/internal/modfetch: move Import and Query to ../modload
Import and Query are about the set of modules viewed
through the modifying lens of the exclude and replace
directives in go.mod. Since package modload is what
manages go.mod, move them there.

Query's 'allowed' argument was meant to capture the
go.mod influence, but it fails to capture replacements,
and at that point we might as well move it where it belongs.

Import is currently ignorant of module replacements
but needs to take them into account.

This CL only moves the code and test. It does not make semantic changes.

Change-Id: I7adde3b0e6076f8ce5719e86c1f771d8dcb4b572
Reviewed-on: https://go-review.googlesource.com/122880
Reviewed-by: Bryan C. Mills <bcmills@google.com>